cid#1532367 Unused value
Within the loop nPrevRowChange is not read after it is set
and nPrevRowChange is always set to nRowChange at the end of
the loop. There is no continue within the loop, so it looks
like there is no point changing nPrevRowChange as it will be
overwritten with nRowChange in any case.

probably this is the case since:

commit cc52d895314dd7b67de916bd90ccbfa098e77419
Date:   Wed Jun 7 16:53:42 2023 +0200

    tdf#155342 sw tracked table column: manage changes

Change-Id: Ib7cf2da23a7553bdeea13862cf4f67248bc829a7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/152994
Tested-by: Jenkins
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>