tdf#133507 sw: hunt for documents with ULSpacing at section end
When a paragraph style has "contextualspacing" with Spacing Above paragraph,
and option "Don't add space between paragraphs of the same style,
the spacing should apply after a section end.

Contextual spacing was officially added to 1.3 spec in 7.0-ish timeframe:
"The space between the paragraphs is zero,
if all of the following conditions hold:
...
 The paragraphs belong to the same content area.
...

Contextual spacing was actually added in LO 3.6,
and it applies as expected when transitioning
from one section to another.

The missing case was when transitioning from a section
back to the main body text.

This was an implementation error. So the question is
whether we add yet another compatibility flag
to handle the old way of doing things,
or just fix it.

I'm leaning towards no compat flag because
-it is a recent official addition to ODF
-it is an implementation error
-it currently is inconsistent - works on section start, but not end
-rarely used: it is primarily an interoperability feature
-rarely encountered: sections aren't really needed in LO.
-no existing unit tests matched the condition.
-interoperability isn't affected - section end == document end.

This patch is hunting to prove some of the above points.
-shouldn't hit any DOC/DOCX/RTF files.
-only expect to find ODT files related to this bug report.

So, I don't think a compat flag is warranted in this case.
It is easy enough for a human to "fix" the document
by removing the undesired "spacing before/after" settings.

make CppunitTest_sw_odfexport2 \
    CPPUNIT_TEST_NAME=tdf133507_contextualSpacingSection

Change-Id: Ia041741bbada8c2f4cd97aa6b7cd98d4d7ffb275
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153033
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>