sw floattable: fix handling of upper margin of anchor for fly intersect
The bugdoc has 2 paragraphs and a shape between them. Word has a small
amount of space between the shape and the 2nd paragraph, Writer has a
large amount of space between the shape and the 2nd paragraph.

It seems what happens is that Writer puts both the text node's own
upper spacing and the previous text node's lower spacing into the upper
spacing of the text frame, so the check added in commit
d07fc485d46f431405a3f6a002f951a08c559677 (tdf#116486 Consider upper
margin in paragraph positioning with flys, 2018-04-10) is not entirely
correct, as it deal with both upper and lower margins, but its intention
was just the upper margins.

Fix the problem by (indirectly) using
GetUpperSpaceAmountConsideredForPrevFrame(), which allows doing the
intersection for the original text node upper spacing but not for the
text node lower spacing. This keeps the original bugdoc working, but
fixes the unexpected vertical space for the new bugdoc.

This is part of the effort so that the original bnc#816603 bugdoc's DOCX
version moves an overlapping floating table from page 4 back to page 3,
like Word does. testTdf122878 is modified to just make sure no overlap
happens, manual testing of the bugdoc shows that our layout still
matches Word.

Change-Id: I4622fb77dc8a52493766e50688ec92065eac65bc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153101
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153118