tdf#133507 sw layout: fix missing contextual spacing at section end
Most of the fix was included in 12f256cd1ae3e625c7b7f48987b7e843046924fa
which was hunting for examples. Bad idea. Merge these two together.

So this patch is just removing the document hunting aspect
and enabling the fix itself. (Yeah, really bad idea.)

When a paragraph style has "contextualspacing" with Spacing Above,
and option "Don't add space between paragraphs of the same style,
the spacing should apply after a section end.

Contextual spacing was officially added to 1.3 spec in 7.0-ish timeframe:
"The space between the paragraphs is zero,
if all of the following conditions hold:
...
    The paragraphs belong to the same content area.
...

Contextual spacing was actually added in LO 3.6,
and it applies as expected when transitioning
from one section to another.

The missing case was when transitioning from a section
back to the main body text.

This was an implementation error. So the question is
whether we add yet another compatibility flag
to handle the old way of doing things,
or just fix it.

I'm leaning towards no compat flag because
-it is a recent official addition to ODF
-it is an implementation error
-it currently is inconsistent - works on section start, but not end
-rarely used: it is primarily an interoperability feature
-rarely encountered: sections aren't really needed in LO.
-no existing unit tests matched the condition.
-interoperability isn't affected - section end == document end.

Miklos agreed that a compat flag is warranted in this case.
It is easy enough for a human to "fix" the document
by removing the undesired "spacing before/after" settings.

make CppunitTest_sw_odfexport2 \
    CPPUNIT_TEST_NAME=tdf133507_contextualSpacingSection

Change-Id: I5db903e0f1ba41aab54f881256c50d3d1927d6d1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153187
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>