tdf#154521 sw navigator: allow to query selected bookmark via UNO
If the hidden title of SwNavigatorPanel was emptied via UNO XPanel
interface, store the name of the selected bookmark there. This allows
to query the selected bookmark using UNO e.g. in add-ons, i.e. to
disambiguate when multiple bookmarks are there on the selected text
range.

Note: this is a workaround because getDialog() of XPanel is not
implemented for SwNavigatorPanel.

Follow-up to commit c4a58634753a84b09f20f7271d6525a6656522d3
"tdf#154545 sw Navigator: select & track nested bookmarks" and
commit 6eb1d540a1e599aa4fe0a321eddb9cc22e0546d3
"tdf#154521 sw Navigator: fix selection change event of bookmark".

Change-Id: I94f79daf59516331155e0b36502821c769771207
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153162
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>