tdf#132061: disable uno:PageColumnType when read-only
When the document is read-only, it should not be able
to be have the column information changed,
so disable that toolboxControl, just like everything else.

It was set to be active in READ-ONLY mode since initial import,
but no indication why it should be different
from the two other page actions added at the same time.

It was added as part of adding the sidebar.
However, I don't see where it would be found in the sidebar.
It is not in the normal menus - just in the notebookbar.

Change-Id: Icf342d5ee51cf8bba296049833e0e520813efb03
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153193
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>