tdf#142833 sd notebookbar: There is no PrintPreview, use WebHtml
Well, that surprised me, but there is no print preview available
in Draw or Impress. They can only do a pathetic
Preview in Web Browser.

AFAIAC, this could be removed since it is so useless.
But that would leave an empty hole under "Print directly"
so for now just substitute something useless for something non-functional.

Change-Id: If7fbb1b549c2b24d2ca6f77f14333811c405e465
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153196
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>