tdf#155827 MCGR: Axial to linear in oox export, apply...
... separately for color and transparency gradient.

The current solution had the change from axial to linear applied to
color and transparency gradient at the same time. That resulted in
wrong gradient if only one of them was axial.

Because changing axial to linear changes the number of stops, it has
to be applied before color and transparency gradients are synchronized.

As now color and transparency are both linear already after they
are synchronized, the switch can be replaced with a simple if.

Change-Id: I75e6b495782bc8a9fe94504426efadec2f60b61f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153197
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
(cherry picked from commit b08ceed6e82e58089e58d15e5ac694b8517349be)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153331
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>