tdf#118335 notebookbar: set visible false for PrintPreview in MenuFile
The label for "Print Preview" comes from ContextLabel in
these menus - and it displays "Close Preview" even if
the print preview is not active.

The PrintPreview buttons in the notebookbar ribbon itself work fine.

So just hide the menu ones for now.
Perhaps that is not good for accessibility,
but I would imagine accessibility people would
stick with the traditional menu anyway,
since the notebookbar is also horrendous
in many other ways for accessiblity.

The user can turn this back on again if they want.
Customize - Notebookbar - File Tab | File | Menu File

Change-Id: I807941650295132c9e15f04094cece26246318d3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153199
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>