tdf#103064 Don't overwrite all formatting with DF on lang change
i.e. right click on the language field in the statusbar, and
choose a language from the popup menu.

We now start with an empty itemset and only put there the
modified items. This in turn makes the spelling restating
hack unnecessary, so move it near its last consumer in
ResetLanguages.

Change-Id: Ie0f97a7fa41dc19b5785a41ee2ef4fb9c814f74a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153252
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>