tdf#103064 Refresh spellcheck marks on default lang change
i.e. Tools > Language > For All Text, or Tools > Options... >
Language Settings > Languages.

The new language was already set as a pool default and picked by
comments, just the spellcheck marks weren't refreshed.

Another issue is that (unlike the Options dialog) menu items for
specific languages (e.g. "English (USA)") also reset language
settings in all paragraph/character styles, as implemented in
tdf#35282, and this needs to be broadcasted as style modification.
However as it turns out, the problem described in tdf#35282 was
also reported in i#102191, for which the OOo devs provided a
different fix (setting the new lang as DF), which was eventually
merged as well as commit 8780cc5c294085fe7379a24eee58ac446cc40d0c
("cws tl84: #i102191# 'set language for all text' now uses
hard language attribute"). So now we have two solutions for
the same problem applied both. Given this, I prefer to instead
revert the commit of tdf#35282, as it seems odd that a menu that
claims to modify text actually modifies arbitrary styles (which
might not even be used in the text).

(fwiw there is also a third possible solution for i#102191/tdf#35282
which wasn't even discussed: Just rename the menu title from
"For All Text" to "Document", as the problem seems to be really
that the menu title sets wrong expectations that it modifies the
text attributes, and not just the document defaults.)

An open question regarding the fix of i#102191 is whether
comments and shapes should also get the default language as DF.
This actually used to work for comments until recently, as we used
to force the document body language on attached comments. But this
is no longer the case, so a different solution would need to be
developed if this is still wanted.

Change-Id: Ib097765e30a7272270737fc61df6d10aa51565fb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153477
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>