tdf#155920: next effect in END mode should no resume show
When starting a presentation and moving over the last slide into the END
mode the next effect should not resume the presentation. Instead it
should exit the presentation. Furthermore, when hitting then the back
key it should show again the last slide (only resuming the presentation,
not directly resuming and moving back to the slide before the last.

Change-Id: I816232725d923601356e0bc8e39421f256d184eb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153302
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>