tdf#155887 Add individual WMF records validation
In previous implementation, there was no proper validation
of the records, by it's size.
This commit adds validation of the WMF records.

It would allow to interrupt reading, if wrong record found
and play only records until it it correct.

It will allow to mimic the behaviour of MS Office:
If the wrong record is detected, the image is displayed till the wrong record.

Currently in LibreOffice, if wrong record is found whole read is interrupted.

Change-Id: I0c82deabcec7a416ca44540e693822f1986437eb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153351
Tested-by: Jenkins
Reviewed-by: Bartosz Kosiorek <gang65@poczta.onet.pl>