tdf#103064 Fix reset to parent style handling
If the style ends up empty, it still needs modification
broadcasting if it used to be non-empty. So let's just
trust the pool notification for when to notify clients.
And given that broadcasting happens now always after
EnsureStyleHierarchy, I assume that the explicit
broadcasting in SetParent is no longer needed.

Also added broadcasting to the undo of a reset. And the
undo actions were combined, so we don't get separate
notifications for each reset attribute.

Change-Id: Ia2895fe346ef337cc0b4fe5dc4275f5b2dc60cd7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153478
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>