tdf#107787 Sync FillGradientStepCount and StepCount
The FillGradientStepCount property of a shape or page background and
the StepCount member of the Gradient2 API struct or nStepCount member
of the basegfx::BGradient class are used parallel and mixed. Therefore
we need to be careful to keep the values in sync as far as possible.

Change-Id: I58ab9654ba0106417794fafe68fb296e66cb3bf5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153714
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>
(cherry picked from commit 48a9ade1dacc63e61cc9a5748f29119d1d01d841)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153764
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>