lok: use correct shell when doing undo
When doing undo we save EditShell reference in the context object.
Later it is used to get cursor we want to move to the place
where modification was done (in sw/source/core/undo/unins.cxx:281).

In LOK case when 2 different sessions were used and undo was done
by user A - then user's B cursor was moved. This fixes that issue.

We use current shell instead of getting it indirectly through
SwDoc::GetEditShell()

Change-Id: I7a10ea98587f48818e8d0aaa9ad739d4eed514f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153461
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>