related tdf#68695 tdf#141136 sc: comment can't resize; avoid context change
In order to avoid having the notebookbar
constantly flipping back to the home tab,
try to avoid useless context switches.
That should also be a nice performance win.

In this case, the context change occurred because handles were enabled,
but that broke in LO 4.1, so just revert that broken OOo fix.

Back in OOo, https://bz.apache.org/ooo/show_bug.cgi?id=33764
added resize handles to a newly inserted comment,
and was added to LO as part of a huge batch.
commit 07b0c01857137de7e4bd02375e7c892d622d270e
Author: Oliver Bolte on Mon Mar 2 14:19:13 2009 +0000
    CWS-TOOLING: integrate CWS dr66

But that didn't allow select and copy, so in LO 4.1
commit 54a1feb9b9bd654774b9aa60cda7ef9a1cd11064
Author: Armin Le Grand on Fri May 10 12:08:40 2013 +0100
    Resolves: #i122142# use simple text hilight frames

Bug 68695 exists to complain about the inability to resize.

Change-Id: If04d718b7de9cff1038bbe9f6372d4859643d3f2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153508
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>