tdf#151548 ContentControls vba: allow search by float #2
I don't know how to figure out where the basic parsing
is mishandling the double-specifier of "#",
so just hack in a re-attempt to find a control with the
name of an unfound index.

This should be fine. If it really was specifying an index,
then it won't get to this code.

If the number is larger than the index and a control exists
with that id, then it almost certainly was intended as a name.
If by some chance it wasn't, then instead of the code failing,
it will actually go ahead with the operation.

Since in practical terms this won't happen,
just go ahead and fix the likely case of a
positive-number-as-a-name control not being found.

Change-Id: I2a34184d86b99a50fdc6b806264f1d5e7794ad48
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153630
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
(cherry picked from commit 609a1567d0e60ca11800df56059b97b6a61ad117)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153605