tdf#125543 Fix window size in GTK3 plugin
When importing a csv file, the size of the main window Calc in windowed mode were not set
because of erroneous check for widget visibility in gtk3.
Now it isn't the visibility of the widget that is checked, but whether the checked widget is a window.
A similar problem was observed in Base

Change-Id: I48f882f407ba17c6c983c5b0939b9ff789ee59f6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153670
Tested-by: Jenkins
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
(cherry picked from commit 431dc963a0d18341c9b4869a2a64009b7ff7f699)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157119
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>