fix cross-compilation with frozen headers
FROZEN wasn't part of the PERMITTED_BUILD_TARGETS in configure.ac, that
in turn means it is not part of BUILD_TYPE when building the crossbuild
build-tools.
That in turn means gb_Helper_optional in external/Module_external.mk
doesn't add it to the external module dirs, and that means the makefiles
from external/frozen aren't read, more specifically there won't be any
gb_UnpackedTarball_UnpackedTarball,frozen definition.

The UnpackedTarball definitions don't error out in this case, and the
gb_Library_use_extenals,module,frozen then doesn't work as expected. It
happily depends on the UnpackedTarball_get_final_target for frozen, but
since there is nothing to unpack (since the makefiles with the
definitions are skipped) that is a noop/it goes straight to just touch
workdir_for_build/UnpackedTarball/frozen.update to fulfill the
dependency without actually extracting any files.

This patch deals with the first problem only, adding it to the
PERMITTED_BUILD_TARGETS will allow the mechanism to work as intended.

Change-Id: Ie9e8ad47ba4c281fb3246daf67450a16f792b908
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153747
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>