tdf#118335 remove "Close Preview" PopupLabel from uno:PrintPreview
The PopupLabel cannot be depended on to only be used
when PrintPreview is already running.
Specifically, in the notebookbar menus,
it always said "Close Preview" even when
the preview was not yet toggled on.

Instead, use uno:ClosePreview in contexts
that only show up when the preview is running.
This has already been done in LO codebase.

Change-Id: I3c743da895f9ac4febc18329f251d283c0cec1cf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153779
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>
Reviewed-by: Justin Luth <jluth@mail.com>