Revert "tdf#118335 notebookbar: set visible false for PrintPreview in MenuFile"
This reverts 24.2 commit 506db23af4da2c5c8775fb0fd6fb9b8668e2fe86,
because the PopupLabel has been removed, so the label showing for Print Preview
is now "Print Preview" instead of "Close Preview".

Because it is a toggle, it will have a checkmark as the picture
when the feature is active (and thus would act as a close preview).

The "close preview" instances were converted to .uno:ClosePreview
to allow the PopupLabel to be removed.

Change-Id: I7c38e863b39d3f3979466a8ab10f29dcda561d22
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/153811
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>